go-cross: Bring go-cross version 1.8
The intel quark family lacks the MMX instruction set, hence go needs
to be patched. Meta-iot2000 only provides these patches
for versions 1.6 and 1.8. Our go-cross provider is meta-resin
which only provides 1.7, so we need to provide 1.8.
Meta-resin is our go-cross provider because it has a
higher layer priority than oe-meta-go.

Signed-off-by: Theodor Gherzan <theodor@resin.io>