local.conf.sample: Delete references to RESIN_CONNECTABLE* variables
Starting with meta-resin 2.14.0, we allow building of one image which
by default is unmanaged. The conversion between managed and unmanaged
will be done through the os-config tool at boot time, based on the
existence of resinio configuration parameters in config.json, so
we remove the references to these obsolete variables from here.

Signed-off-by: Florin Sarbu <florin@balena.io>